+++ CHANGELOG.md
@@ -6,6 +6,7 @@ versioning is done in a continuous fashion without worries of breaking changes.
 
 ## patches
 
+- `systemd`: keep track of tasks or at least todos of the slack app 2026-02-09
 - `zip`: add zipping and unzipping programs to tom for vim viewings 2026-02-07
 - `nvim`: support highlighting an html templating engine called eta 2026-02-07
 - `zsh`: alias zips to worktrees matching alphabetic alias creation 2026-02-07